[Update] Rename component
