+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import HomePage from './pages/HomePage/index.jsx';
+import IntroPage from './pages/IntroPage/index.jsx';
 import RegisterPage from './pages/RegisterPage/index.jsx';
 import LoginPage from './pages/LoginPage/index.jsx';
 
@@ -8,7 +8,7 @@ function App() {
   return (
     <div>
       <Routes>
-        <Route path="/" element={<HomePage />} />
+        <Route path="/" element={<IntroPage />} />
         <Route path="/register" element={<RegisterPage/>} />  
         <Route path="/login" element={<LoginPage/>} />  
       </Routes>
