+++ src/pages/IntroPage/index.jsx
@@ -1,22 +1,22 @@
 import React from "react";
-import Home from "../../components/partials/Home";
+import Intro from "../../components/partials/Intro";
 import Footer from "../../components/layouts/Footer";
-import Header from "../../components/partials/Home/Header";
+import Header from "../../components/partials/Intro/Header";
 import topImage from "../../assets/img/main.png";
 import styles from "./style.module.scss";
 
-const HomePage = () => {
+const IntroPage = () => {
   return (
-    <div className={styles.homePage}>
+    <div className={styles.introPage}>
       <div className={styles.topSection}>
         <img src={topImage} alt="persfash top" className={styles.topImage} />
       </div>
       <div className={styles.space}></div>
       <Header />
-      <Home />
+      <Intro />
       <Footer />
     </div>
   );
 };
 
-export default HomePage;
+export default IntroPage;
