+++ src/pages/IntroPage/style.module.scss
@@ -1,4 +1,4 @@
-.homePage {
+.introPage {
   position: relative;
   height: 100%; /* Adjust height as necessary */
 }
